move functions to utils to undo circular dependencies
