+++ nimare/generate.py
@@ -7,7 +7,8 @@ import numpy as np
 from .dataset import Dataset
 from .io import convert_neurovault_to_dataset
 from .meta.utils import compute_ale_ma, get_ale_kernel
-from .transforms import mm2vox, transform_images, vox2mm
+from .transforms import transform_images
+from .utils import vox2mm, mm2vox
 
 # defaults for creating a neurovault dataset
 NEUROVAULT_IDS = (8836, 8838, 8893, 8895, 8892, 8891, 8962, 8894, 8956, 8854, 9000)
